requirements.txt
python-dotenv
python-telegram-bot
groq
openai
chatgpt-md-converter